Prepare 'SweMini' for .Net Core.
