+++ SwissEphNet.sln
@@ -17,8 +17,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Programs", "Programs", "{DA
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SwissEphNet", "SwissEphNet\SwissEphNet.csproj", "{8895A5B1-E188-4926-B204-87B921EE040C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SweMini", "Programs\SweMini\SweMini.csproj", "{D2043376-EA04-4677-803E-698CB384D361}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SweTest", "Programs\SweTest\SweTest.csproj", "{29AABD40-C169-40B9-8A84-2CFA070CE9E5}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SweWin", "Programs\SweWin\SweWin.csproj", "{D1AE1640-218C-4CCB-BD09-EADCAE96335A}"
@@ -31,6 +29,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "CI", "CI", "{D0592369-8075-
 		SwissEphNet\SwissEphNet.nuspec = SwissEphNet\SwissEphNet.nuspec
 	EndProjectSection
 EndProject
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "SweMini", "Programs\SweMini\SweMini.xproj", "{67BE4773-CC1F-4C5B-A3A6-38A163C9768C}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -41,10 +41,6 @@ Global
 		{8895A5B1-E188-4926-B204-87B921EE040C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8895A5B1-E188-4926-B204-87B921EE040C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8895A5B1-E188-4926-B204-87B921EE040C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D2043376-EA04-4677-803E-698CB384D361}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D2043376-EA04-4677-803E-698CB384D361}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D2043376-EA04-4677-803E-698CB384D361}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D2043376-EA04-4677-803E-698CB384D361}.Release|Any CPU.Build.0 = Release|Any CPU
 		{29AABD40-C169-40B9-8A84-2CFA070CE9E5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{29AABD40-C169-40B9-8A84-2CFA070CE9E5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{29AABD40-C169-40B9-8A84-2CFA070CE9E5}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -57,14 +53,18 @@ Global
 		{93ADF5AE-AF28-4BF2-9918-6F442A8F6ABE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{93ADF5AE-AF28-4BF2-9918-6F442A8F6ABE}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{93ADF5AE-AF28-4BF2-9918-6F442A8F6ABE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{67BE4773-CC1F-4C5B-A3A6-38A163C9768C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{67BE4773-CC1F-4C5B-A3A6-38A163C9768C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{67BE4773-CC1F-4C5B-A3A6-38A163C9768C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{67BE4773-CC1F-4C5B-A3A6-38A163C9768C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{D2043376-EA04-4677-803E-698CB384D361} = {DA9EB626-FC40-46C3-A44D-3EE76D044CC8}
 		{29AABD40-C169-40B9-8A84-2CFA070CE9E5} = {DA9EB626-FC40-46C3-A44D-3EE76D044CC8}
 		{D1AE1640-218C-4CCB-BD09-EADCAE96335A} = {DA9EB626-FC40-46C3-A44D-3EE76D044CC8}
 		{93ADF5AE-AF28-4BF2-9918-6F442A8F6ABE} = {6D5C2702-2FF1-42A8-B86D-E5E5EAE41F66}
+		{67BE4773-CC1F-4C5B-A3A6-38A163C9768C} = {DA9EB626-FC40-46C3-A44D-3EE76D044CC8}
 	EndGlobalSection
 EndGlobal
